requirements.txt
# Pipeline
prefect
librosa
numpy
torch
torchaudio

# API
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
boto3

# Shared
psycopg[binary]
python-dotenv